python_service/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
protobuf==4.21.12
torch==2.0.0
torchvision==0.15.1
Pillow==9.2.0
numpy<2
